Added trigger none test
