+++ src/AzurePipelinesToGitHubActionsConverter.Tests/ConditionTests.cs
@@ -148,58 +148,6 @@ namespace AzurePipelinesToGitHubActionsConverter.Tests
         }
 
         [TestMethod]
-        public void NestedStringTest()
-        {
-            //Arrange
-            string text = "(this is a (sample) string with (some special words. (another one)))";
-
-            //Act
-            List<string> results = ConditionsProcessing.FindBracketedContentsInString(text);
-
-            //Assert
-            Assert.AreNotEqual(null, results);
-            Assert.AreEqual(4, results.Count);
-        }
-
-        [TestMethod]
-        public void NestedString2Test()
-        {
-            //Arrange
-            string text = "not(contains('ABCDE', 'BCD'))";
-
-            //Act
-            List<string> results = ConditionsProcessing.FindBracketedContentsInString(text);
-
-            //Assert
-            Assert.AreNotEqual(null, results);
-            Assert.AreEqual(2, results.Count);
-            Assert.AreEqual("'ABCDE', 'BCD'", results[0]);
-            Assert.AreEqual("contains('ABCDE', 'BCD')", results[1]);
-        }
-
-        [TestMethod]
-        public void MultilineBracketsTest()
-        {
-            //Arrange
-            string text = @"
-and(
-succeeded(),
-or(
-    eq(variables['Build.SourceBranch'], 'refs/heads/master'),
-    startsWith(variables['Build.SourceBranch'], 'refs/tags/')
-),
-contains(variables['System.TeamFoundationCollectionUri'], 'dsccommunity')
-)";
-
-            //Act
-            List<string> results = ConditionsProcessing.FindBracketedContentsInString(text);
-
-            //Assert
-            Assert.AreNotEqual(null, results);
-            Assert.AreEqual(6, results.Count);
-        }
-
-        [TestMethod]
         public void MultilineConditionTest()
         {
             //Arrange
@@ -250,6 +198,138 @@ startsWith(variables['Build.SourceBranch'], 'refs/tags/')
             //Assert
             string expected = "startsWith(github.ref, 'refs/tags/')";
             Assert.AreEqual(expected, result);
+        }
+
+        [TestMethod]
+        public void NestedBracketsStringTest()
+        {
+            //Arrange
+            string text = "(this is a (sample) string with (some special words. (another one)))";
+
+            //Act
+            List<string> results = ConditionsProcessing.FindBracketedContentsInString(text);
+
+            //Assert
+            Assert.AreNotEqual(null, results);
+            Assert.AreEqual(4, results.Count);
+        }
+
+        [TestMethod]
+        public void NestedBracketsString2Test()
+        {
+            //Arrange
+            string text = "not(contains('ABCDE', 'BCD'))";
+
+            //Act
+            List<string> results = ConditionsProcessing.FindBracketedContentsInString(text);
+
+            //Assert
+            Assert.AreNotEqual(null, results);
+            Assert.AreEqual(2, results.Count);
+            Assert.AreEqual("'ABCDE', 'BCD'", results[0]);
+            Assert.AreEqual("contains('ABCDE', 'BCD')", results[1]);
+        }
+
+        [TestMethod]
+        public void MultilineBracketsTest()
+        {
+            //Arrange
+            string text = @"
+and(
+succeeded(),
+or(
+    eq(variables['Build.SourceBranch'], 'refs/heads/master'),
+    startsWith(variables['Build.SourceBranch'], 'refs/tags/')
+),
+contains(variables['System.TeamFoundationCollectionUri'], 'dsccommunity')
+)";
+
+            //Act
+            List<string> results = ConditionsProcessing.FindBracketedContentsInString(text);
+
+            //Assert
+            Assert.AreNotEqual(null, results);
+            Assert.AreEqual(6, results.Count);
+        }
+
+        [TestMethod]
+        public void SimpleTwoSplitTest()
+        {
+            //Arrange
+            string condition = "'ABCDE', 'BCD'";
+
+            //Act
+            List<string> results = ConditionsProcessing.SplitContents(condition);
+
+            //Assert
+            Assert.AreEqual(2, results.Count);
+
+        }
+
+        [TestMethod]
+        public void TrickSingleSplitWithBracketsTest()
+        {
+            //Arrange
+            string condition = "contains('ABCDE', 'BCD')";
+
+            //Act
+            List<string> results = ConditionsProcessing.SplitContents(condition);
+
+            //Assert
+            Assert.AreEqual(1, results.Count);
+        }
+
+        [TestMethod]
+        public void ComplexDoubleNestedBracketsSplitTest()
+        {
+            //Arrange
+            string condition = "('ABCDE', 'BCD'), ne(0, 1)";
+
+            //Act
+            List<string> results = ConditionsProcessing.SplitContents(condition);
+
+            //Assert
+            Assert.AreEqual(2, results.Count);
+        }
+
+        [TestMethod]
+        public void SimpleThreeSplitTest()
+        {
+            //Arrange
+            string condition = "succeeded(), variables['Build.SourceBranch'], 'refs/heads/master'";
+
+            //Act
+            List<string> results = ConditionsProcessing.SplitContents(condition);
+
+            //Assert
+            Assert.AreEqual(3, results.Count);
+        }
+
+
+        [TestMethod]
+        public void ComplexDoubleSplitWithNesterBracketTest()
+        {
+            //Arrange
+            string text = @"succeeded1(),or(succeeded2(),succeeded3())";
+
+            //Act
+            List<string> results = ConditionsProcessing.SplitContents(text);
+
+            //Assert
+            Assert.AreEqual(2, results.Count);
+        }
+
+        [TestMethod]
+        public void ComplexTripleSplitWithDoubleNestedBracketsTest()
+        {
+            //Arrange
+            string text = @"succeeded(),eq('ABCDE', 'BCD'), ne(0, 1)";
+
+            //Act
+            List<string> results = ConditionsProcessing.SplitContents(text);
+
+            //Assert
+            Assert.AreEqual(3, results.Count);
         }
 
     }
